Minor changes and future plans
Thought about the next steps.

Start with building logic for how game differs in first and second innings.

We need initial impact stats for bowlers and batsmen, which will vary significantly based on their play during the match.

We need to pass extra information as to who is bowling and who is batting. Also we need to take care of handling extras.

How will match situation vs player's impact vary. If Dhoni comes into bat when the team needs 30 of 10, will the odds be drastically in the opponents favour or not.